requirements.txt
base==0.0.0
gpiod==2.4.1
gpiodevice==0.0.5
importlib_resources==6.5.2
inky==2.3.0
IT8951 @ git+https://github.com/GregDMeyer/IT8951.git@9f136139378f74e17d9972d7165dc6ae53a2568e
Jetson.GPIO==2.1.12
numpy==2.4.3
packaging==26.0
pillow==12.1.1
pygame==2.6.1
RPi.GPIO==0.7.1
smbus2==0.6.0
spidev==3.8
toml==0.10.2